moved `decode_audio_data` in `BaseAudioContext` + merged `audio_buffer` in `buffer`
