+++ examples/bench.rs
@@ -76,5 +76,5 @@ fn main() {
 
     let mut context = context;
     let output = context.start_rendering();
-    assert_eq!(output.sample_len(), len);
+    assert_eq!(output.length(), len);
 }
