+++ src/lib.rs
@@ -49,9 +49,6 @@ pub const RENDER_QUANTUM_SIZE: usize = 128;
 /// Maximum number of channels for audio processing
 pub const MAX_CHANNELS: usize = 32;
 
-// experimental
-pub mod audio_buffer;
-
 pub mod buffer;
 pub mod context;
 pub mod control;
