+++ tests/offline.rs
@@ -27,15 +27,15 @@ fn test_offline_render() {
 
     let output = context.start_rendering();
     assert_eq!(output.number_of_channels(), 2);
-    assert_eq!(output.sample_len(), LENGTH);
+    assert_eq!(output.length(), LENGTH);
 
     assert_float_eq!(
-        output.channel_data(0).as_slice(),
+        output.get_channel_data(0),
         &[-2.; LENGTH][..],
         abs_all <= 0.
     );
     assert_float_eq!(
-        output.channel_data(1).as_slice(),
+        output.get_channel_data(1),
         &[-2.; LENGTH][..],
         abs_all <= 0.
     );
@@ -62,9 +62,9 @@ fn test_start_stop() {
 
     let output = context.start_rendering();
     assert_eq!(output.number_of_channels(), 1);
-    assert_eq!(output.sample_len(), RENDER_QUANTUM_SIZE * 4);
+    assert_eq!(output.length(), RENDER_QUANTUM_SIZE * 4);
 
-    let channel_data = output.channel_data(0).as_slice();
+    let channel_data = output.get_channel_data(0);
 
     // one chunk of silence, two chunks of signal, one chunk of silence
     let mut expected = vec![0.; RENDER_QUANTUM_SIZE];
@@ -91,9 +91,9 @@ fn test_delayed_constant_source() {
 
     let output = context.start_rendering();
     assert_eq!(output.number_of_channels(), 1);
-    assert_eq!(output.sample_len(), RENDER_QUANTUM_SIZE * 4);
+    assert_eq!(output.length(), RENDER_QUANTUM_SIZE * 4);
 
-    let channel_data = output.channel_data(0).as_slice();
+    let channel_data = output.get_channel_data(0);
 
     // two chunks of silence, two chunks of signal
     let mut expected = vec![0.; 2 * RENDER_QUANTUM_SIZE];
@@ -129,9 +129,9 @@ fn test_audio_param_graph() {
 
     let output = context.start_rendering();
     assert_eq!(output.number_of_channels(), 1);
-    assert_eq!(output.sample_len(), RENDER_QUANTUM_SIZE);
+    assert_eq!(output.length(), RENDER_QUANTUM_SIZE);
 
-    let channel_data = output.channel_data(0).as_slice();
+    let channel_data = output.get_channel_data(0);
 
     // expect output = 0.8 (input) * ( 0.5 (intrinsic gain) + 0.4 (via 2 constant source input) )
     let expected = vec![0.8 * 0.9; RENDER_QUANTUM_SIZE];
@@ -186,7 +186,7 @@ fn test_cycle() {
     let output = context.start_rendering();
     // cycle should be muted, and other source should be processed
     assert_float_eq!(
-        output.channel_data(0).as_slice(),
+        output.get_channel_data(0),
         &[2.; RENDER_QUANTUM_SIZE][..],
         abs_all <= 0.
     );
@@ -218,17 +218,17 @@ fn test_cycle_breaker() {
 
     // not muted, and positive feedback cycle
     assert_float_eq!(
-        output.channel_data(0).as_slice()[..RENDER_QUANTUM_SIZE],
+        output.get_channel_data(0)[..RENDER_QUANTUM_SIZE],
         &[1.; RENDER_QUANTUM_SIZE][..],
         abs_all <= 0.
     );
     assert_float_eq!(
-        output.channel_data(0).as_slice()[RENDER_QUANTUM_SIZE..2 * RENDER_QUANTUM_SIZE],
+        output.get_channel_data(0)[RENDER_QUANTUM_SIZE..2 * RENDER_QUANTUM_SIZE],
         &[2.; RENDER_QUANTUM_SIZE][..],
         abs_all <= 0.
     );
     assert_float_eq!(
-        output.channel_data(0).as_slice()[2 * RENDER_QUANTUM_SIZE..3 * RENDER_QUANTUM_SIZE],
+        output.get_channel_data(0)[2 * RENDER_QUANTUM_SIZE..3 * RENDER_QUANTUM_SIZE],
         &[3.; RENDER_QUANTUM_SIZE][..],
         abs_all <= 0.
     );
